moved gower distance test classes
