+++ corehunter-base/src/test/resources/Datos_11_25_2013-.csv
@@ -1,0 +1,1 @@
+Fruit_diam,Seed_wght,nr_flow_axila,Fruit_wght,Plant_hght,Fruit_length,Seed_nr,Corola_color,Anther_color,Corola_spot_color,Fruit_epid,Fruit_form,Flower_position,Corola_form,10,20,0.3,1,69.4,40.8,37.3,32.1,4,2,0,2,2,5,111,9.5,0.2,2,6.7,33.7,10.7,17.5,4,2,0,2,2,5,118,24.9,0.4,1.3,122.4,38.3,51,42.1,4,2,0,2,2,5,141,16.3,0.4,1.7,28,43.8,47.1,15.8,4,2,0,2,2,5,143,14.3,0.4,2,35.5,42.7,55.3,43.3,4,2,0,2,2,5,150,29.2,0.4,2,104.1,55.6,51.8,84.6,4,2,0,2,2,5,151,21.9,0.4,2,85.8,50.6,57.2,46.7,4,2,0,2,2,5,154,27,0.4,2,79.4,56.6,55,54.4,4,2,0,2,2,5,167,14.9,0.5,1.7,30.4,62.7,38.3,17,4,2,0,2,2,5,168,14.7,0.3,2,21.3,56.7,21.9,43.1,4,2,0,2,2,5,185,19.2,0.5,1.3,41.8,62.2,27.3,43.7,4,2,0,2,2,5,1147,26.5,0.7,1,73.5,51.9,23.5,93.7,1,2,0,2,2,5,1149,22.4,0.6,1,55.8,54,21.7,83.5,1,2,0,2,2,5,1167,21.7,0.5,1,63.3,63.4,36.9,101.1,1,2,0,2,2,5,1175,18.3,0.6,1,146.5,58.6,82.2,82.3,1,2,0,2,2,5,1179,11.8,0.5,1,38.8,69.8,68.1,84.7,1,2,0,2,2,5,1181,20,0.6,1,51.4,62.8,35.8,105.5,1,2,1,2,2,5,1205,33,0.5,1.3,198.7,59.1,63.8,77.3,1,2,1,2,2,5,1212,11.3,0.3,2,15.9,40.6,19.7,22.1,1,2,1,2,2,5,1215,11.1,0.4,2,22.1,49.6,44,43.2,1,2,1,2,2,5,1216,4.9,0.4,2,5.1,65.4,19.5,16.7,1,2,1,2,2,5,2221,11.8,0.3,2,13.5,63.1,18.3,29.1,1,2,1,2,2,6,2224,17.7,0.5,2.3,36.7,51,48.1,47.7,1,3,1,1,2,6,2225,13.1,0.3,2,12.9,34.7,12.8,14.6,1,3,1,1,2,6,2233,28,0.3,2.7,86.7,40.9,46.7,80.7,1,3,1,1,2,6,2239,10.3,0.4,1.3,22.9,58,38.2,20.5,1,3,1,1,3,6,2242,4.4,0.4,1.3,5.7,71.6,25.3,9.7,1,3,1,1,3,6,2246,15.2,0.3,2,33.9,46.5,42.2,38.9,1,3,1,1,3,6,2250,16.3,0.4,2,41.7,58.5,48,37.5,1,3,1,1,3,6,2252,18.6,0.4,1.3,32.3,51.3,20.6,48.5,1,3,1,1,3,6,2268,10.6,0.4,1.7,12.4,59.4,42.6,22.4,4,3,1,1,3,6,2275,18.8,0.4,2,102.9,58.9,82.9,75.3,4,3,1,1,3,6,3298,11.1,0.3,2,6.8,56.4,11,18.9,4,3,1,1,3,6,3301,15.3,0.3,2,17.1,48,31.5,34.7,4,3,1,1,3,7,3340,12.7,0.4,1,31.9,53.3,43.9,53.8,4,3,1,1,3,7,3347,21.5,0.6,1,160.3,66.1,97.1,118.4,4,3,1,1,3,7,3363,27.3,0.7,1,127.4,64.1,48.2,58,4,3,1,1,3,7,3406,22.5,0.3,1.3,28.5,54.5,16.6,46.1,4,3,1,1,3,7,3407,22.1,0.5,1.3,53.7,54,29.3,53.9,4,3,1,1,3,7,3
